UnitTests for UserRestController (not all working)
